requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
altair==5.4.0
    # via streamlit
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.4.0
    # via streamlit
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
clip==0.2.0
    # via -r requirements.in
faiss-cpu==1.8.0.post1
    # via -r requirements.in
filelock==3.15.4
    # via
    #   torch
    #   triton
fsspec==2024.6.1
    # via torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
idna==3.7
    # via requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
    #   torch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
narwhals==1.3.0
    # via altair
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   faiss-cpu
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.20
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
packaging==24.1
    # via
    #   altair
    #   faiss-cpu
    #   streamlit
pandas==2.2.2
    # via streamlit
pillow==10.4.0
    # via streamlit
protobuf==4.25.4
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via streamlit
rich==13.7.1
    # via streamlit
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.37.1
    # via -r requirements.in
sympy==1.13.2
    # via torch
tenacity==8.5.0
    # via streamlit
toml==0.10.2
    # via streamlit
torch==2.4.0
    # via -r requirements.in
tornado==6.4.1
    # via streamlit
triton==3.0.0
    # via torch
typing-extensions==4.12.2
    # via
    #   altair
    #   streamlit
    #   torch
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.2
    # via streamlit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
